Performance modifications.
add a method(named "onOriginalBitmapLoaded") in Callback to notfiy user that bitmap is loaded from internet or diskcard.

It's called in background  thread.In the method of onOriginalBitmapLoaded,you can do transformation or resize bitmap(I create an background bitmap based on original one and merge them together).After that the bitmap will be cached in memory and diskcard.So if you do some transformation in this method,you may  don't have to call Request#transformation or resize,or so on.